+++ lib/assets/javascripts/new-dashboard/components/Wizard/Header.vue
@@ -7,7 +7,7 @@
           v-for="(stepName, index) in stepNames" :key="stepName">
           <a href="">
             <span class="breadcrumbs__checkpoint" :class="[isCurrentStep (currentStep, index) ? 'current' : '']">
-              <span class="breadcrumbs__text">{{ index + 1 }} - {{ stepName }}</span>
+              <span class="breadcrumbs__text">{{ stepName }}</span>
             </span>
           </a>
         </li>
@@ -35,15 +35,19 @@ export default {
 <style scoped lang="scss">
 @import 'new-dashboard/styles/variables';
 
-$timeline__border-width: 4px;
+$timeline__border-width: 1px;
+$timeline__border-color: #DDD;
+$timeline__border-transition: 0.25s;
+$bullet__transition: 0.12s;
+$bullet__border-color: #D3E6FA;
+$transition__timing-function: cubic-bezier(0.4, 0.01, 0.165, 0.99);
 
 .header {
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  padding: 4em 0 2em;
-  border-bottom: 1px solid $grey;
+  padding-top: 2.406em;
   background: $white;
 }
 
@@ -59,10 +63,10 @@ $timeline__border-width: 4px;
   &__item {
     position: relative;
     width: 100%;
-    height: 50px;
+    height: 46px;
 
     &:not(:last-child) {
-      border-top: $timeline__border-width solid $primary-color--soft;
+      border-top: $timeline__border-width solid $primary-color;
     }
 
     &:last-child {
@@ -70,15 +74,46 @@ $timeline__border-width: 4px;
       border-top: $timeline__border-width solid transparent;
     }
 
+    &::after {
+      content: '';
+      display: block;
+      position: absolute;
+      top: -$timeline__border-width;
+      right: 0;
+      width: 0%;
+      transition: width $timeline__border-transition $transition__timing-function;
+      border-top: $timeline__border-width solid $timeline__border-color;
+    }
+
     &.current {
-      border-color: $grey;
+      &::after {
+        content: '';
+        display: block;
+        position: absolute;
+        top: -$timeline__border-width;
+        right: 0;
+        width: 100%;
+        transition: width $timeline__border-transition $transition__timing-function;
+        border-top: $timeline__border-width solid $timeline__border-color;
+      }
 
       &:last-child {
         border-color: transparent;
       }
 
       & ~ .breadcrumbs__item {
-        border-color: $grey;
+        border-color: $timeline__border-color;
+
+        &::after {
+          content: '';
+          display: block;
+          position: absolute;
+          top: -$timeline__border-width;
+          right: 0;
+          width: 100%;
+          transition: none;
+          border-top: $timeline__border-width solid $timeline__border-color;
+        }
 
         &:last-child {
           border-color: transparent;
@@ -86,12 +121,15 @@ $timeline__border-width: 4px;
       }
 
       & ~ .breadcrumbs__item .breadcrumbs__checkpoint {
-        top: -10px;
-        left: -10px;
-        width: 16px;
-        height: 16px;
-        border: $timeline__border-width solid $grey;
-        background: $white;
+        top: -4px;
+        left: -4px;
+        width: 7px;
+        height: 7px;
+        background: $timeline__border-color;
+      }
+
+      & ~ .breadcrumbs__item .breadcrumbs__text {
+        color: $text-secondary-color;
       }
     }
   }
@@ -100,7 +138,7 @@ $timeline__border-width: 4px;
     position: absolute;
     top: 30px;
     transform: translateX(-50%);
-    color: $text-secondary-color;
+    color: $text-color;
     font-size: 0.75em;
     white-space: nowrap;
   }
@@ -108,23 +146,24 @@ $timeline__border-width: 4px;
   &__checkpoint {
     display: block;
     position: absolute;
-    top: -9px;
-    left: -9px;
-    width: 14px;
-    height: 14px;
+    z-index: 1;
+    top: -4px;
+    left: -4px;
+    width: 7px;
+    height: 7px;
+    transition: all $bullet__transition $transition__timing-function;
     border-radius: 50%;
-    background: $primary-color--soft;
+    background: $primary-color;
 
     &.current {
-      top: -6px;
-      left: -6px;
-      width: 8px;
-      height: 8px;
-      background: $primary-color;
-      box-shadow: 0 0 0 12px rgba($primary-color, 0.2);
+      transition: all $bullet__transition $transition__timing-function;
+      transition-delay: $timeline__border-transition;
+      box-shadow: 0 0 0 6px $bullet__border-color;
 
       .breadcrumbs__text {
-        color: $primary-color;
+        transition: all $bullet__transition $transition__timing-function;
+        color: $text-color;
+        font-weight: 600;
       }
     }
   }